+++ CHANGELOG.md
@@ -12,8 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Spawn multiple commands in parallel with `--threads`.
 - Add options for command output handling:
   - `--no-header`
-  - `--output` with `inherit` and `null`
+  - `--output` with `inherit`, `line-prefix` and `null`
   - `--result` with `always`, `never` and `non-zero`
+  - `--line-prefix` as a shortcut for `--no-header --output=line-prefix`
   - `--only-result` as a shortcut for `--no-header --output=null`
   - `--quiet` as a shortcut for `--no-header --output=null --result=never`